requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.2
python-dotenv==1.0.1
sse-starlette==2.1.0
